+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+* Update dependency on `ember-flexberry-data` to version `3.6.0-beta.0`.
 
 ## [2.6.0-beta.0] - 2020-10-06
 ### Added